Refactor metal buffer concepts
