+++ candle-metal-kernels/src/metal/mod.rs
@@ -1,6 +1,9 @@
+pub mod buffer;
 pub mod device;
 pub mod library;
 pub mod metal_utils;
+
+pub use buffer::*;
 pub use device::*;
 pub use library::*;
 pub use metal_utils::*;
